Add support for PXE root pivoting.
In some environments, it is desirable to export the kernel, loader.conf, etc. from a shared NFS server with a standard DAC-based security policy and the root filesystem (which contains things like setuid binaries) from another server. This commit adds support for such a two-stage NFS boot by allowing pxeboot to 'pivot' from the boot filesystem specified by DHCP to a root filesystem specified in loader.conf.

The technique is to make the PXE 'device' listen for the DIOCGFLUSH ioctl, which is sent just before the loader boots the kernel. On receiving this ioctl, if vfs.root.mountfrom specifies a different NFS location from the current root, PXE 'pivots' to the new VFS root and updates the appropriate environmental variables for the early kernel boot code, which speaks NFS but not the mount protocol.