Allow rtld direct-exec to take a file descriptor.
Summary:
When executing rtld directly, allow the value of argv[0] to specify a
file descriptor of the binary to execute. If this fails, fall back to
the previous behaviour of treating argv[0] as a path.

Given the newfound ability of rtld to be directly executed, this approach can be
used as an alternative to creating an ffexecve(2) system call.

This diff has two commits in it: one to rename parse_libdir->parse_integer
and the other to actually do the exec-an-fd work.

Reviewers: kib, rwatson

Subscribers: brooks, emaste

Differential Revision: https://reviews.freebsd.org/D10751