Make "compile one file" requirement more explicit.
When creating a C/C++ Toolchain, pass in a function that can be used to
compile a single source file to a single object file. This allows the
code creating the Toolchain (e.g., `gcc-like.fab`) to specify how a
single object file should be compiled while allowing the common
Toolchain logic to handle dealing with multiple files.